+++ packages/core/styles/src/packages/composite/sidebar/sidebar.sprinkle.css.ts
@@ -16,6 +16,9 @@ const defineProperties = defineRainbowProperties({
     lg: {
       "@media": mediaQueries.lg(),
     },
+    xl: {
+      "@media": mediaQueries.xl(),
+    },
   },
   defaultCondition: "xs",
   dynamicProperties: {