+++ packages/react/src/atomic/Box/src/box.stories.tsx
@@ -96,6 +96,18 @@ export default {
     backgroundImage: {
       control: { type: "text" },
     },
+    backgroundPosition: {
+      control: { type: "text" },
+    },
+    backgroundBlendMode: {
+      control: { type: "text" },
+    },
+    backgroundRepeat: {
+      control: { type: "text" },
+    },
+    backgroundSize: {
+      control: { type: "text" },
+    },
     minWidth: {
       control: { type: "text" },
     },
@@ -125,26 +137,31 @@ box.args = {
 export const responsive = Template.bind({});
 responsive.args = {
   height: {
+    xl: "6rem",
     lg: "4rem",
     md: "3rem",
     xs: "2rem",
   },
   width: {
+    xl: "12rem",
     lg: "10rem",
     md: "8rem",
     xs: "2rem",
   },
   backgroundColor: {
+    xl: "warning.interactive",
     lg: "success.interactive",
     md: "danger.interactive",
     xs: "primary.interactive",
   },
   paddingX: {
+    xl: "8",
     lg: "6",
     md: "4",
     xs: "none",
   },
   borderRadius: {
+    xl: "1rem",
     lg: "0.5rem",
     md: "0.2rem",
     xs: "100%",