+++ packages/react/src/composite/Table/src/table.stories.tsx
@@ -28,10 +28,10 @@ base.args = {
   children: (
     <>
       <Table.Head>
-        <Table.Cell>Header 1</Table.Cell>
-        <Table.Cell>Header 2</Table.Cell>
-        <Table.Cell>Header 3</Table.Cell>
-        <Table.Cell>Header 4</Table.Cell>
+        <Table.Cell as="th">Header 1</Table.Cell>
+        <Table.Cell as="th">Header 2</Table.Cell>
+        <Table.Cell as="th">Header 3</Table.Cell>
+        <Table.Cell as="th">Header 4</Table.Cell>
       </Table.Head>
       <Table.Body>
         {Array.from({ length: 10 }, (_, i) => (
@@ -52,9 +52,13 @@ withCustomCellWidth.args = {
   children: (
     <>
       <Table.Head>
-        <Table.Cell width="180px">This column has fixed width</Table.Cell>
-        <Table.Cell>This column has auto width</Table.Cell>
-        <Table.Cell width="180px">This column has fixed width</Table.Cell>
+        <Table.Cell as="th" width="180px">
+          This column has fixed width
+        </Table.Cell>
+        <Table.Cell as="th">This column has auto width</Table.Cell>
+        <Table.Cell as="th" width="180px">
+          This column has fixed width
+        </Table.Cell>
       </Table.Head>
       <Table.Body>
         {Array.from({ length: 10 }, (_, i) => (
@@ -74,9 +78,15 @@ withNoPadding.args = {
   children: (
     <>
       <Table.Head>
-        <Table.Cell padding="none">Header 1</Table.Cell>
-        <Table.Cell padding="none">Header 2</Table.Cell>
-        <Table.Cell padding="none">Header 3</Table.Cell>
+        <Table.Cell as="th" padding="none">
+          Header 1
+        </Table.Cell>
+        <Table.Cell as="th" padding="none">
+          Header 2
+        </Table.Cell>
+        <Table.Cell as="th" padding="none">
+          Header 3
+        </Table.Cell>
       </Table.Head>
       <Table.Body>
         {Array.from({ length: 10 }, (_, i) => (