+++ packages/react/src/atomic/Link/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 The Link component allows us to navigate to external addresses.
 
-## 2023-02-24 `3.5.1`
+## 2023-02-24 `3.6.0`
+
+### 🎉 New features
+
+- Replaced `size` property with `fontSize`. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `lineHeight` property to component API. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
 
 ### 💡 Others